resmon: Introduce the BPF interface
To be able to analyze the EMADs, they need to be transported to the
user-space. In this patch, resmon installs a BPF program hooked to the
devlink_hwmsg tracepoint. This program does a shallow dissection of the
EMADs just to discover what register they carry, and if an interesting one,
it pushes it to the ring buffer, whence the demon picks it up for further
processing.

The ring buffer is represented by a file descriptor that it is possible to
poll on. Introduce a resmon_back API to yield the descriptor, so that the
resmon_d_loop() can incorporate and poll on it. Also introduce an activity
callback, invoked when the poll indicates the backend FD has data to read.
This then dispatches the read EMADs to the EMAD dissection introduced in
the previous patches.

The HW backend is made accessible and resmon default at this point.
Invoking the mock backend can be done through the "resmon start mode mock".

Signed-off-by: Petr Machata <petrm@nvidia.com>