Move common helpers to a dedicated directory
The following commits will add resmon, a tool for tracking resource usage
on a Spectrum switch. resmon consists of a daemon, a CLI interface, a BPF
program, and is more complex than the other libbpf-based tools. It will
therefore be in a separete directory, with its own Makefile. However it
will still want to reuse the common helpers.

In this patch, move the common helpers into a dedicated top-level
directory, add a Makefile, and have it build libcommon.a with the helper
objects. Propagate the configuration to the downsteam Makefile.

Signed-off-by: Petr Machata <petrm@nvidia.com>